+++ util/random_utils.cpp
@@ -13,7 +13,7 @@ float RandomBetween(float min, float max)
     static bool randInitialized = false;
     if (!randInitialized)
     {
-        srand(static_cast<unsigned int>(time(NULL)));
+        srand(time(NULL));
         randInitialized = true;
     }
